# CT: Add hash256