--- a/go/ct/st/state_test.go
+++ b/go/ct/st/state_test.go
@@ -52,7 +52,7 @@ func getNewFilledState() *State {
 	s.LastCallReturnData = NewBytes([]byte{1})
 	s.HasSelfDestructed = true
 	s.SelfDestructedJournal = []SelfDestructEntry{{vm.Address{1}, vm.Address{2}}}
-	s.RecentBlockHashes = [256]vm.Hash{{0x01}}
+	s.RecentBlockHashes = NewImmutableHashArray(vm.Hash{0x01})
 	return s
 }
 
@@ -164,7 +164,7 @@ func getTestChanges() map[string]testStruct {
 			"Different has-self-destructed journal entry",
 		},
 		"block_number_hashes": {func(state *State) {
-			state.RecentBlockHashes = [256]vm.Hash{{0x02}}
+			state.RecentBlockHashes = NewImmutableHashArray(vm.Hash{0x02})
 		},
 			"Different block number hash at index 0: 0200000000000000000000000000000000000000000000000000000000000000 vs 0100000000000000000000000000000000000000000000000000000000000000",
 		},
@@ -451,7 +451,7 @@ func TestState_PrinterMemorySize(t *testing.T) {
 func TestState_PrinterRecentBlockHashes(t *testing.T) {
 	s := NewState(NewCode([]byte{byte(BLOCKHASH)}))
 	s.Stack.Push(NewU256(0))
-	s.RecentBlockHashes = [256]vm.Hash{{0x01}}
+	s.RecentBlockHashes = NewImmutableHashArray(vm.Hash{0x01})
 
 	r := regexp.MustCompile(`Hash of block 0: 0x([0-9a-fA-F]+)`) // \[([0-9a-f]+)\]
 	str := s.String()
@@ -484,7 +484,7 @@ func TestState_DiffMatch(t *testing.T) {
 	s1.LastCallReturnData = NewBytes([]byte{1})
 	s1.HasSelfDestructed = true
 	s1.SelfDestructedJournal = []SelfDestructEntry{{vm.Address{0x01}, vm.Address{0x01}}}
-	s1.RecentBlockHashes = [256]vm.Hash{{0x01}}
+	s1.RecentBlockHashes = NewImmutableHashArray(vm.Hash{0x01})
 
 	s2 := NewState(NewCode([]byte{byte(PUSH2), 7, 4, byte(ADD), byte(STOP)}))
 	s2.Status = Running
@@ -503,7 +503,7 @@ func TestState_DiffMatch(t *testing.T) {
 	s2.LastCallReturnData = NewBytes([]byte{1})
 	s2.HasSelfDestructed = true
 	s2.SelfDestructedJournal = []SelfDestructEntry{{vm.Address{0x01}, vm.Address{0x01}}}
-	s2.RecentBlockHashes = [256]vm.Hash{{0x01}}
+	s2.RecentBlockHashes = NewImmutableHashArray(vm.Hash{0x01})
 
 	diffs := s1.Diff(s2)
 
@@ -535,7 +535,7 @@ func TestState_DiffMismatch(t *testing.T) {
 	s1.LastCallReturnData = NewBytes([]byte{1})
 	s1.HasSelfDestructed = true
 	s1.SelfDestructedJournal = []SelfDestructEntry{{vm.Address{0x01}, vm.Address{0x01}}}
-	s1.RecentBlockHashes = [256]vm.Hash{{0x01}}
+	s1.RecentBlockHashes = NewImmutableHashArray(vm.Hash{0x01})
 
 	s2 := NewState(NewCode([]byte{byte(PUSH2), 7, 5, byte(ADD)}))
 	s2.Status = Running
@@ -554,7 +554,7 @@ func TestState_DiffMismatch(t *testing.T) {
 	s2.LastCallReturnData = NewBytes([]byte{249})
 	s2.HasSelfDestructed = false
 	s2.SelfDestructedJournal = []SelfDestructEntry{{vm.Address{0xf3}, vm.Address{0xf3}}}
-	s2.RecentBlockHashes = [256]vm.Hash{{0xf2}}
+	s2.RecentBlockHashes = NewImmutableHashArray(vm.Hash{0xf2})
 
 	diffs := s1.Diff(s2)
 
@@ -786,7 +786,7 @@ func TestState_EqualityConsidersRelevantFieldsDependingOnStatus(t *testing.T) {
 			relevantFor: allButFailed,
 		},
 		"block_number_hashes": {
-			modify:      func(s *State) { s.RecentBlockHashes = [256]vm.Hash{{0xf2}} },
+			modify:      func(s *State) { s.RecentBlockHashes = NewImmutableHashArray(vm.Hash{0xf2}) },
 			relevantFor: allButFailed,
 		},
 	}