--- a/go/vm/interpreter.go
+++ b/go/vm/interpreter.go
@@ -10,9 +10,7 @@
 
 package vm
 
-import (
-	"fmt"
-)
+import "fmt"
 
 //go:generate mockgen -source interpreter.go -destination interpreter_mock.go -package vm