This PR adds an immutable `hash256` type, analogous to `Bytes` to improve performance on state operations.

For instance in `Clone`
```
facu@:~/Tosca$ benchstat oldHashClone.log newHashClone.log
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Tosca/go/ct/st
cpu: AMD Ryzen 7 PRO 7840U w/ Radeon 780M Graphics
                    │ oldHashClone.log │          newHashClone.log           │
                    │      sec/op      │   sec/op     vs base                │
State_CloneState-16        9.839µ ± 5%   7.832µ ± 4%  -20.40% (p=0.000 n=25)

                    │ oldHashClone.log │           newHashClone.log           │
                    │       B/op       │     B/op      vs base                │
State_CloneState-16       41.67Ki ± 0%   32.91Ki ± 0%  -21.02% (p=0.000 n=25)

                    │ oldHashClone.log │        newHashClone.log        │
                    │    allocs/op     │ allocs/op   vs base            │
State_CloneState-16         13.00 ± 0%   13.00 ± 0%  ~ (p=1.000 n=25) ¹
¹ all samples are equal
```

And for `Eq`
```
facu@:~/Tosca$ benchstat oldHashEq.log newHashEq.log
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Tosca/go/ct/st
cpu: AMD Ryzen 7 PRO 7840U w/ Radeon 780M Graphics
                 │ oldHashEq.log │           newHashEq.log            │
                 │    sec/op     │   sec/op     vs base               │
State_EqState-16    1005.0n ± 3%   788.2n ± 6%  -21.57% (p=0.000 n=25)

                 │ oldHashEq.log │         newHashEq.log          │
                 │     B/op      │    B/op     vs base            │
State_EqState-16      0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=25) ¹
¹ all samples are equal

                 │ oldHashEq.log │         newHashEq.log          │
                 │   allocs/op   │ allocs/op   vs base            │
State_EqState-16      0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=25) ¹
¹ all samples are equal
```