@LuisPH3 2024-06-12
I had an idea about this, but I don't know if it is possible or clean.

If the conditions do not contain any constraint for the `recentBlockHashes`  you may leave it nill, fill it to zeroes when translating into the VM state.
 
The blockhash history is a read only construct from the point of view of the VM, I am not sure if we even need to compare the state after calling the VM. 

What do you guys think?